+++ tests/test_crypt.py
@@ -23,6 +23,7 @@ except NameError:
   # For Python3 (though importlib should be used, silly 3.3).
   from imp import reload
 
+from oauth2client import _helpers
 from oauth2client.client import HAS_OPENSSL
 from oauth2client.client import SignedJwtAssertionCredentials
 from oauth2client import crypt
@@ -46,16 +47,24 @@ class Test_pkcs12_key_as_pem(unittest.TestCase):
         scope='read+write',
         sub='joe@example.org')
 
-  def test_succeeds(self):
+  def _succeeds_helper(self, password=None):
     self.assertEqual(True, HAS_OPENSSL)
 
     credentials = self._make_signed_jwt_creds()
-    pem_contents = crypt.pkcs12_key_as_pem(credentials.private_key,
-                                           credentials.private_key_password)
+    if password is None:
+      password = credentials.private_key_password
+    pem_contents = crypt.pkcs12_key_as_pem(credentials.private_key, password)
     pkcs12_key_as_pem = datafile('pem_from_pkcs12.pem')
-    pkcs12_key_as_pem = crypt._parse_pem_key(pkcs12_key_as_pem)
+    pkcs12_key_as_pem = _helpers._parse_pem_key(pkcs12_key_as_pem)
     alternate_pem = datafile('pem_from_pkcs12_alternate.pem')
     self.assertTrue(pem_contents in [pkcs12_key_as_pem, alternate_pem])
+
+  def test_succeeds(self):
+    self._succeeds_helper()
+
+  def test_succeeds_with_unicode_password(self):
+    password = u'notasecret'
+    self._succeeds_helper(password)
 
   def test_without_openssl(self):
     import imp
